backend/requirements.txt
Django>=5.2,<5.3
wagtail>=7.2,<7.3
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.9,<5.0

# Celery для асинхронных задач
celery>=5.3,<6.0
redis>=5.0,<6.0

# Агрегация контента
pytrends>=4.9,<5.0
feedparser>=6.0,<7.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
lxml>=5.0,<6.0

# Telegram интеграция
telethon>=1.34,<2.0

# AI интеграции
# openai>=1.0,<2.0  # для будущего использования OpenAI API
# stability-sdk>=0.8,<1.0  # для будущего использования Stability AI
gradio_client>=1.4,<2.0

# Дополнительные утилиты
Pillow>=10.0,<11.0
python-dateutil>=2.8,<3.0
python-dotenv>=1.0,<2.0
